Added 'what is this file' at top
This is an example template for anyone wishing to add their hardware to the Haiku forum manually; this file was last updated 3 Jan 2020. **Please** do not create an edit or pull request to contribute; to add hardware, please use this template with the 'computers compatible (v3)' thread on Haiku or the Google Form. For edit requests, please post your concerns on the Haiku forum. Thank you!